backend/requirements.txt
Django==6.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.0
pandas==2.1.3
requests==2.31.0
reportlab==4.0.7
python-dotenv==1.0.0
